Refactor resolvers into more granular modules.
