+++ packages/openneuro-server/graphql/resolvers/description.js
@@ -1,0 +1,6 @@
+/**
+ * Resolver implementation for dataset_description.json
+ */
+import * as datalad from '../../datalad/description.js'
+
+export const description = obj => datalad.description(obj.id)
