+++ packages/openneuro-server/graphql/resolvers/summary.js
@@ -9,3 +9,21 @@ export const summary = dataset => {
     datasetId: dataset.id,
   })
 }
+
+/**
+ * Save summary data returned by the datalad service
+ *
+ * Returns the saved summary if successful
+ */
+export const updateSummary = (obj, args) => {
+  const summaries = mongo.collections.crn.summaries
+  return summaries
+    .update(
+      { id: args.summary.id, datasetId: args.summary.datasetId },
+      args.summary,
+      {
+        upsert: true,
+      },
+    )
+    .then(() => args.summary)
+}
