+++ packages/openneuro-server/graphql/resolvers/validation.js
@@ -2,24 +2,6 @@ import mongo from '../../libs/mongo.js'
 import pubsub from '../pubsub'
 
 /**
- * Save summary data returned by the datalad service
- *
- * Returns the saved summary if successful
- */
-export const updateSummary = (obj, args) => {
-  const summaries = mongo.collections.crn.summaries
-  return summaries
-    .update(
-      { id: args.summary.id, datasetId: args.summary.datasetId },
-      args.summary,
-      {
-        upsert: true,
-      },
-    )
-    .then(() => args.summary)
-}
-
-/**
  * Save issues data returned by the datalad service
  *
  * Returns only a boolean if successful or not
